Rename the interactions.ts file to match symbol
