+++ src/lib/InteractionManager.ts
